requirements.txt
praw
python-telegram-bot